src/test/java/com/iteso/nintendo/GrassTest.java
package com.iteso.nintendo;


import org.junit.Before;
import org.junit.Test;

import static org.junit.Assert.assertEquals;
import static org.mockito.Mockito.mock;
import static org.mockito.Mockito.when;

public class GrassTest {

    Grass grass;

    Attack mockedAttack;

    @Before
    public void setUp() {
        grass = new Grass();

        mockedAttack = mock(Attack.class);

    }

    @Test
    public void isSuperEffective() {
        when(mockedAttack.getAttackType()).thenReturn("fire");

        assertEquals(2, grass.defends(mockedAttack), 0.0f);

    }
    @Test
    public void isUnneffective() {
       when(mockedAttack.getAttackType()).thenReturn("grass");

        assertEquals(0.5, grass.defends(mockedAttack), 0.0f);

    }
    @Test
    public void isNeutral() {
        when(mockedAttack.getAttackType()).thenReturn("normal");

        assertEquals(1, grass.defends(mockedAttack), 0.0f);

    }

}

